feature: WEOS-1343 Create middleware for handling JWT in the incoming Authorization header -Moved authorization middleware to standard controller file as well as test to that test file
