+++ controllers/rest/middleware_initialize_test.go
@@ -1,11 +1,7 @@
 package rest_test
 
 import (
-	"bytes"
-	"encoding/json"
 	"io/ioutil"
-	"net/http"
-	"net/http/httptest"
 	"os"
 	"regexp"
 	"strings"
@@ -103,62 +99,6 @@ func TestXRemove(t *testing.T) {
 
 		if authorTable.HasField("Email") {
 			t.Errorf("expected the struct to not have field '%s'", "Email")
-		}
-	})
-}
-
-func TestAuthenticateMiddleware(t *testing.T) {
-	//instantiate api
-	api, err := rest.New("./fixtures/blog-security.yaml")
-	if err != nil {
-		t.Fatalf("unexpected error loading api '%s'", err)
-	}
-	err = api.Initialize(context.TODO())
-	if err != nil {
-		t.Fatalf("un expected error initializing api '%s'", err)
-	}
-	e := api.EchoInstance()
-
-	t.Run("no jwt token added when required", func(t *testing.T) {
-		description := "testing 1st blog description"
-		mockBlog := &TestBlog{Description: &description}
-		reqBytes, err := json.Marshal(mockBlog)
-		if err != nil {
-			t.Fatalf("error setting up request %s", err)
-		}
-		body := bytes.NewReader(reqBytes)
-		resp := httptest.NewRecorder()
-		req := httptest.NewRequest(http.MethodPost, "/blogs", body)
-		req.Header.Set("Content-Type", "application/json")
-		e.ServeHTTP(resp, req)
-		if resp.Result().StatusCode != http.StatusUnauthorized {
-			t.Errorf("expected the response code to be %d, got %d", http.StatusUnauthorized, resp.Result().StatusCode)
-		}
-	})
-	t.Run("security parameter array is empty", func(t *testing.T) {
-		resp := httptest.NewRecorder()
-		req := httptest.NewRequest(http.MethodGet, "/health", nil)
-		e.ServeHTTP(resp, req)
-		if resp.Result().StatusCode != http.StatusOK {
-			t.Errorf("expected the response code to be %d, got %d", http.StatusOK, resp.Result().StatusCode)
-		}
-	})
-	t.Run("jwt token added", func(t *testing.T) {
-		description := "testing 1st blog description"
-		mockBlog := &TestBlog{Description: &description}
-		reqBytes, err := json.Marshal(mockBlog)
-		if err != nil {
-			t.Fatalf("error setting up request %s", err)
-		}
-		token := "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJKb2huIiwibmFtZSI6IkpvaG4gRG9lIiwiZW1haWwiOiJqRG9lQGdtYWlsLmNvbSIsImlhdCI6MTY0NDg4Nzc2OSwiZXhwIjoxNjQ0ODg4NzY5fQ.sBL03kXCIbjzD5MdzCRb71g8LLQhgr9R7a0-3cJxySw"
-		body := bytes.NewReader(reqBytes)
-		resp := httptest.NewRecorder()
-		req := httptest.NewRequest(http.MethodPost, "/blogs", body)
-		req.Header.Set("Content-Type", "application/json")
-		req.Header.Set("Authorization", "Bearer "+token)
-		e.ServeHTTP(resp, req)
-		if resp.Result().StatusCode != http.StatusCreated {
-			t.Errorf("expected the response code to be %d, got %d", http.StatusCreated, resp.Result().StatusCode)
 		}
 	})
 }
